+++ gea/libs/gea_math/src/gea/mth_geometry/polygon2.cpp
@@ -25,7 +25,7 @@ const float polygon2::circumference() const
     const pointset_type &points = m_container.m_points;
 
     float circumference = 0.0f;
-    for (int i = 0, e = points.size(); i < e; ++i) {
+    for (size_t i = 0, e = points.size(); i < e; ++i) {
         circumference += points[i].to(points[(i + 1) % e]).length();
     }
     return circumference;
@@ -38,10 +38,10 @@ const float polygon2::area() const
     const pointset_type &points = m_container.m_points;
 
     float area = 0.0f;
-    for (int i = 0, e = points.size(); i < e; ++i) {
+    for (size_t i = 0, e = points.size(); i < e; ++i) {
         area += determinant(points[i], points[(i + 1) % e]);
     }
-    return fabs(area * 0.5f);
+    return fabsf(area * 0.5f);
 }
 
 // ------------------------------------------------------------------------- //
